Iq parser could be better broken up so that the behaviour belonging to events and ozone is separate.
Moved ozone specific functionality to a separate parser and have the processing of events under an event processor.

No tests were removed, all correctly re-factored.

There is now a clear distinction between ozone specification and what happens to Connfu::Events. This will help if the underlying ozone implement was to be replaced.