Once an event has been handled, execution should always return to the end of the latest handle_event call
Due to the funky way continuations affect the call stack, if a handle_event call triggers a #continue, it previously could exit from the previous call to handle_event!  Sounds crazy, but it's true.  Here's an attempt at a simplified example:

  handle_event(:wait)
  puts "Exited #handle_event(:wait)"
  handle_event(:continue)
  puts "Exited #handle_event(:continue)"

In this code, assuming handle_event(:wait) causes a wait, and handle_event(:continue) causes a continue, the continue would previously return from the handle_event(:wait) call, for the second time.  So the output would be:

  Exited #handle_event(:wait)
  Exited #handle_event(:wait)
  Exited #handle_event(:continue)

By storing the latest call to #handle_event in a second continuation, we can ensure that we always exit from the latest call, resetting the call stack to normal after the event has been handled.  The output following this commit should therefore be:

  Exited #handle_event(:wait)
  Exited #handle_event(:continue)