Ignore further commands if the call has already been hung up.
We introduced a very simple 'finished' state into the handler instance, and send_command will not actually send anything if the handler is in that state. On receiving a hangup presence, after firing any `on_hangup` behaviour the handler enters that state.

Finally, after any event is handled, all handler instances in the finished state are removed by the `EventProcessor`.