It should be possible to change behaviour based on who is calling.
In its simplest implementation, this can be done by exposing the "From" header of the offer. I have done this by adding information to the event and yield that, although we may not actually wish to expose the `Event` object to the script developer.

I have elected to use the "From" header, rather than the `from` attribute of the `offer`, as it seemed to contain more information (like an extended email from, rather than just the sip endpoint). However, this could be wrong.

I also had to allow the `run` method to be called with arguments (since the `on :offer` block becomes that method body).