Fix record_for to work with Tropo fix for issue #105.
Now we get a stop recording event from the server when the "max-length" period expires, so we make the "record_for" method wait for a stop recording event. This means we no longer need to use a sleep statement in conjunction with "record_for".

Part of the fix for Tropo has changed the recording URI to be sent as an attribute on a separate "recording" element as opposed to being on the "stop" element itself.

We've updated an extended the examples to demonstrate the new behaviour.