There's no need for Resque when running everything in a single process.
By adding a level of indirection (in Connfu::Queue), we can decouple the implementation of the queueing mechanism from its interface. And by providing an in-process alternative to Resque (Connfu::Queue::InProcess), there is no need to use Resque when running everything in a single process, like in the examples.