Implicitly hangup calls once the offer handler has finished running
In trying to write tests for this, we found the use of any_instance_of in spec helper made it hard to write expectations, (such as method ordering).  To get around we have stubbed build_instance.  Maybe this could be used more widely, as any_instance is generally a code smell.