Move responsibility for mapping events to handlers to the handlers themselves.
Rather than having a map from call id to call handler, each handler is now asked whether it can handle the event.  The first handler that accepts the event is given it.  This allows handlers claim events based on criteria other than their initial call_id.  An example where this will be useful is when initiating a join, the result can only be matched by the blather-id, not the call id.