A first stab at getting transfer using an underlying nested join command working.
We've gone for a *very* naive implementation which ignores Joined events and just waits for a Hangup event for the original call id.

We found that we needed to use the "bridge" media mode in order to get audio working. It seems that this also fixed the problem of calling iptel.org sip addresses[1].

Known limitations :-
* Only a single end-point is currently supported until we establish that our prism/tropo build can support join to multiple endpoints.
* When the transfer_using_join method returns, the call has ended and no more commands can be issued for that call - despite the fact that the structure of the DSL allows you to add other commands.
* We haven't considered anything but the happy path.

[1] https://github.com/tropo/tropo2/issues/124