Fixed the behaviour that was introduced in b4d871dd943e2dcb6099199c7139a487854f9e21.
The aforementioned commit added the ability to find a suitable call handler based on the call_id *or* the IQ ID of the previously issued command.  Unfortunately, the parser hadn't been updated to initialize the events with the IQ ID.  We updated the tests to use the parser to create the events which forced us to fix the implementation.