It's useful to be able to determine the XMPP ID of the last sent command in tests.
This is important when trying to test handlers which must match using command IDs rather than call IDs, e.g. when the call ID is not yet known.

For this to be deterministic, the XMPP stanza for a command is now only generated once.

We've also changed the TestConnection implementation of `send_command` to return the command ID like the real implementation does.