Combining the dial and outgoing call handling will make establishing outgoing calls easier.
Openvoice needs to be able to establish between two external endpoints. This is performed by first dialing the caller, and then joining that call with a new one to the ultimate recipient.

Previously, the application would have to queue a dial job to be performed by Connfu. This only contains information about the first leg of the overall call (i.e. connfu -> caller). Even if the application made the recipient details available via some external data store (i.e. the DB), the Connfu worker would not necessarily be able to distinguish which recipient should be used for this call; when it finds out about the outgoing call, it only knows who the caller is, and would have to make a "best guess" about their intentions.

By being able to queue an entire "outgoing call" job, which keeps the initial dial and subsequent join together, we can avoid this ambiguity.

We are considering changing *all* dials to work this way, and removing the `on :outgoing_call` syntax. There appear to have been some stories and examples which try to use dial multiple times, but we no longer believe that this functionality is driven by a need for Openvoice. We believe it is possible to address this need using the new syntax, should it ever actually be required.