The handler instance send valid commands without knowing the call & client JIDs.
It only knows these values once a ringing presence is received.

Technically we could construct them as soon as we know the call ID (i.e. at the same time we trigger `on_start`), but it seems more robust to extract it from a presence sent by the server. Besides, you can't meaningfully send a command until the call is ringing anyway.