We should only be able to set the call_id, call_jid & client_jid once on a handler instance.
In commit d1f0678348a3495bd65dd0130d3f4dd04d067f93, receipt of the Ringing event would set the call and client JIDs. However, it is possible that a second Ringing event could be handled by the same instance (for example, when joining with another call). This would overwrite the original values. We cannot demonstrate that this is a problem yet, but it seems likely that it would be.

To be safe, we now only allow these handler instance variables to be set once.