Consolidate the scattered sleeps into an explanatory method.
Basically, we need to give Tropo a bit of time once a call is answered to get itself together, otherwise commands we send cause it to break. Hopefully this can all be removed in the near future, but in the meantime it seems to make sense to call it automatically.