We need to customise the call handler based on the call we're dealing with.
We want to be able to hang up the caller/callee when the other end hangs up.  To do this we need to be able to change the behaviour of the handler depending on the call we're dealing with.  Specifically, we only want to evaluate the on_answer block when we receive the answer event for the call to the caller, and we want to customise the behaviour in on_hangup depending on whether the caller or callee ended the call.  At the moment we can't hang up the callee when the caller hangs up as we can't specify the call_id that the hangup command should execute against.