Simpler, more idiomatic Ruby.
It's a bit crazy to include a module which has actually hidden the methods it wants to make available on instances in another module. I probably did this, so I blame myself. I can understand the motivation to gather some of these methods together, but this is not the way.