Made the end_to_end script an actual spec.
We're feeling the pain of our examples deteriorating.  This is the first step to converting the examples into repeatable tests to exercise the system.

We had a problem running the end_to_end_spec when it ran as part of all the specs (unit, functional and integration).  Although we didn't investigate the problem we found that splitting the rake test tasks fixed it.