This is a spiked approach at modularising how multiple calls are handled.
By 'modularised', I don't mean creating Ruby modules, although I've done that quite a bit. Rather than perhaps while a "handler" can deal with incoming calls on its own, it can create handler-esque objects to delegate the events for outgoing calls to.

This commit creates an internal instance within a handler for each outgoing call. The call instance and the overall handler share quite a bit of behaviour, including responding to events and sending commands back to the server, and so this functionality has been extracted into modules for re-use. I suspect that ultimately there's a better refactoring that doesn't rely so much on modules, but this is progress for now.

My hope is that this direction will eventually transform into a more object-oriented representation of calls during a handler's existence. For example, wouldn't it be nice to be able to do something like:

    call1 = OutgoingCall.new # some params
    call2 = OutgoingCall.new # some other params
    call1.say "OK, join you"
    call2.say "You're about to be joined!"
    call1.join(call1)

Of course, all of this can be hidden behind a DSL at any point. Anyway, maybe that's a bit far-fetched for the moment.