Parse a component hangup presence iq
An example of getting this type of Hangup would be when a call is hung up midway through a say component.

For now we are parsing this into the generic Connfu::Event::Hangup as we believe there is no need to create any distinction between a component hangup and a end of call hangup.