We need to be able to timeout when waiting for events.
We're using EventMachine.add_timer which provides a single-shot mechanism for executing a block after a given elapsed time.

When the timeout occurs execution will continue with the `wait_for` method returning `nil`.

We need this in Openvoice because we will need to implement round-robin dialling ourselves; Tropo/Rayo does not provide a built-in mechanism for timing-out dialled calls that have not answered.