Since 14d6cda4493b82e4308c7601be886e55e8e64ff6 we always get a result from the call to `wait`.
There's no need to check the value of `result` since `wait` now always returns an event - either one of the expected events or a timeout event.

Before the commit mentioned above, I think we were only cancelling the timer when it had not yet fired. But since that commit, we always remove it. I'm not sure the latter was intentional, but I think it's probably sensible anyway as long as it doesn't upset event machine.