Tidied/fixed the test that ensures we don't send multiple hangup commands.
The previous incarnation of the test wasn't doing what was intended.  It was stuck waiting for a result from the answer command because the call_jid of the handler didn't match the call_jid of the answer result iq.  In addition, the expectation that hangup shouldn't be called *at all* was incorrect: we actually expect it to be called *once*.